hacks-backend/requirements.txt
django
djangorestframework
python-dotenv
djangorestframework-jsonapi
django-extensions
django-filter
authlib
requests
django-vite
django-cors-headers